requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.25.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
respx>=0.20.0
paramiko>=3.3.0